Move dismissMention to mentions.ts
