+++ front/lib/api/assistant/conversation/mentions.ts
@@ -1047,3 +1047,181 @@ export async function validateUserMention(
 
   return new Ok(undefined);
 }
+
+export async function dismissMention(
+  auth: Authenticator,
+  {
+    messageId,
+    conversationId,
+    type,
+    id,
+  }: {
+    messageId: string;
+    conversationId: string;
+    type: "user" | "agent";
+    id: string;
+  }
+): Promise<Result<void, APIErrorWithStatusCode>> {
+  const conversationRes = await getConversation(auth, conversationId);
+  if (conversationRes.isErr()) {
+    return new Err({
+      status_code: 404,
+      api_error: {
+        type: "conversation_not_found",
+        message: "Conversation not found",
+      },
+    });
+  }
+
+  const conversation = conversationRes.value;
+
+  // Verify the message exists
+  const message = conversation.content.flat().find((m) => m.sId === messageId);
+
+  if (!message) {
+    return new Err({
+      status_code: 404,
+      api_error: {
+        type: "message_not_found",
+        message: "Message not found",
+      },
+    });
+  }
+
+  if (isUserMessageType(message)) {
+    // Verify user is authorized to edit the message by checking the message user.
+    if (message.user && message.user.id !== auth.getNonNullableUser().id) {
+      return new Err({
+        status_code: 403,
+        api_error: {
+          type: "invalid_request_error",
+          message: "User is not authorized to dismiss this mention",
+        },
+      });
+    }
+  } else if (isAgentMessageType(message)) {
+    // Verify user is authorized to edit the message by going back to the user message.
+    const { userMessageUserId } = await getUserMessageIdFromMessageId(auth, {
+      messageId,
+    });
+    if (
+      userMessageUserId !== null &&
+      userMessageUserId !== auth.getNonNullableUser().id
+    ) {
+      return new Err({
+        status_code: 403,
+        api_error: {
+          type: "invalid_request_error",
+          message: "User is not authorized to dismiss this mention",
+        },
+      });
+    }
+  } else if (isContentFragmentType(message)) {
+    return new Err({
+      status_code: 400,
+      api_error: {
+        type: "invalid_request_error",
+        message: "Invalid message type",
+      },
+    });
+  } else {
+    assertNever(message);
+  }
+
+  const updatedMessages: {
+    userMessages: UserMessageType[];
+    agentMessages: AgentMessageType[];
+  } = {
+    userMessages: [],
+    agentMessages: [],
+  };
+
+  // For user mentions, convert sId to database ID
+  let userIdForQuery: number | undefined;
+  if (type === "user") {
+    const user = await getUserForWorkspace(auth, {
+      userId: id,
+    });
+    if (!user) {
+      return new Err({
+        status_code: 404,
+        api_error: {
+          type: "user_not_found",
+          message: "User not found",
+        },
+      });
+    }
+    userIdForQuery = user.id;
+  }
+
+  const predicate = (m: RichMentionWithStatus) =>
+    (m.status === "agent_restricted_by_space_usage" ||
+      m.status === "user_restricted_by_conversation_access") &&
+    m.type === type &&
+    m.id === id;
+
+  // Find all restricted mentions for the same user/agent on conversation messages latest versions.
+  for (const messageVersions of conversation.content) {
+    const latestMessage = messageVersions[messageVersions.length - 1];
+
+    if (
+      latestMessage.visibility !== "deleted" &&
+      !isContentFragmentType(latestMessage) &&
+      latestMessage.richMentions.some(predicate)
+    ) {
+      const mentionModel = await MentionModel.findOne({
+        where: {
+          workspaceId: conversation.owner.id,
+          messageId: latestMessage.id,
+          ...(type === "user"
+            ? { userId: userIdForQuery }
+            : { agentConfigurationId: id }),
+        },
+      });
+      if (!mentionModel) {
+        continue;
+      }
+      await mentionModel.update({ dismissed: true });
+      const newRichMentions = latestMessage.richMentions.map((m) =>
+        predicate(m)
+          ? {
+              ...m,
+              dismissed: true,
+            }
+          : m
+      );
+      if (isUserMessageType(latestMessage)) {
+        updatedMessages.userMessages.push({
+          ...latestMessage,
+          richMentions: newRichMentions,
+          mentions: newRichMentions.map(toMentionType),
+        });
+      } else if (isAgentMessageType(latestMessage)) {
+        updatedMessages.agentMessages.push({
+          ...latestMessage,
+          richMentions: newRichMentions,
+        });
+      }
+    }
+  }
+
+  for (const userMessage of updatedMessages.userMessages) {
+    await publishMessageEventsOnMessagePostOrEdit(
+      conversation,
+      {
+        ...userMessage,
+        contentFragments: getRelatedContentFragments(conversation, userMessage),
+      },
+      []
+    );
+  }
+
+  if (updatedMessages.agentMessages.length > 0) {
+    await publishAgentMessagesEvents(
+      conversation,
+      updatedMessages.agentMessages
+    );
+  }
+
+  return new Ok(undefined);
+}
